compiles on mac

+++ src/VertexArrayAttribute.cpp
@@ -69,9 +69,9 @@ PyObject * MGLVertexArrayAttribute_bind(MGLVertexArrayAttribute * self, PyObject
 	gl.BindBuffer(GL_ARRAY_BUFFER, buffer->buffer_obj);
 
 	if (self->normalizable) {
-		((gl_attribute_normal_ptr_proc)self->gl_attrib_ptr_proc)(self->location, self->row_length, self->scalar_type, false, stride, (void *)offset);
+		((gl_attribute_normal_ptr_proc)self->gl_attrib_ptr_proc)(self->location, self->row_length, self->scalar_type, false, stride, (void *)(long long)offset);
 	} else {
-		((gl_attribute_ptr_proc)self->gl_attrib_ptr_proc)(self->location, self->row_length, self->scalar_type, stride, (void *)offset);
+		((gl_attribute_ptr_proc)self->gl_attrib_ptr_proc)(self->location, self->row_length, self->scalar_type, stride, (void *)(long long)offset);
 	}
 	
 	gl.VertexAttribDivisor(self->location, divisor);
